Create ShadowView.py
It is dangerous to have a module called `ui` and a file called `UI.py`.  Python can keep these things separate but programmers may get confused.  Once ShadowView is working properly, you should be able to remove UI.py and UI.pyui.

Class names like `ShadowView` should be in CamelCase to help programmers understand that they are classes.

The variable name `i` should be reserved for integer values and not for Images or ImageViews.